Folder refactor Data analisis
